Fix: Pass filenames to ts_capture and related fns
I refactored this away before because I thought it would make testing
easier if I didn't have to work with files.

That made me realize that captures need access to filenames to
integrate correctly with telescope.

So I rewrote it back to the original code mostly.
Working with js files in the filetree rather than code samples embedded
in lua is not at all a problem either